# api/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Sentiment Analyzer API",
        "description": "API documentation for Sentiment Analyzer",
        "contact": {
            "name": "Денис Зямаев",
            "email": "code4un@yandex.ru"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000",
            "description": "Localhost"
        }
    ],
    "paths": {
        "/api/auth/signup": {
            "post": {
                "tags": [
                    "Authentication"
                ],
                "summary": "New user registration",
                "description": "Create user and send confirmation email.",
                "operationId": "ef1f8c61940a64c01bc2b178c28e921f",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name",
                                    "email",
                                    "password",
                                    "password_confirmation"
                                ],
                                "properties": {
                                    "name": {
                                        "description": "Имя пользователя",
                                        "type": "string",
                                        "example": "Денис",
                                        "maxLength": 255,
                                        "minLength": 3
                                    },
                                    "email": {
                                        "description": "Уникальный email адрес",
                                        "type": "string",
                                        "format": "email",
                                        "example": "test@yandex.ru"
                                    },
                                    "password": {
                                        "description": "Пароль (мин. 8 символов, буквы и цифры)",
                                        "type": "string",
                                        "format": "password",
                                        "example": "password123",
                                        "minLength": 8
                                    },
                                    "password_confirmation": {
                                        "description": "Подтверждение пароля",
                                        "type": "string",
                                        "format": "password",
                                        "example": "password123"
                                    },
                                    "remember": {
                                        "description": "Флаг \"Запомнить меня\" для продления жизни сессии",
                                        "type": "boolean",
                                        "example": true
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Успешная регистрация. Сессионная кука установлена.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Регистрация успешна. Письмо с подтверждением отправлено на вашу почту."
                                        },
                                        "user": {
                                            "properties": {
                                                "id": {
                                                    "type": "integer",
                                                    "example": 1
                                                },
                                                "name": {
                                                    "type": "string",
                                                    "example": "Денис"
                                                },
                                                "email": {
                                                    "type": "string",
                                                    "example": "test@yandex.ru"
                                                },
                                                "role": {
                                                    "type": "string",
                                                    "example": "user",
                                                    "enum": [
                                                        "user",
                                                        "admin"
                                                    ]
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Ошибка валидации данных (например, email уже занят или пароль слишком простой)",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "The given data was invalid."
                                        },
                                        "errors": {
                                            "description": "Словарь ошибок валидации, где ключ — имя поля, а значение — массив сообщений",
                                            "type": "object",
                                            "additionalProperties": {
                                                "type": "array",
                                                "items": {
                                                    "type": "string",
                                                    "example": "The email has already been taken."
                                                }
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "429": {
                        "description": "Слишком много попыток регистрации (Rate Limit)",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Too Many Attempts."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        }
    },
    "components": {
        "securitySchemes": {
            "sanctum": {
                "type": "http",
                "description": "Enter your token: Bearer {token} (for mobile clients). For SPA (Nuxt.js) auth is via HttpOnly cookies automatically.",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
    "tags": [
        {
            "name": "Authentication",
            "description": "Authentication routes"
        }
    ]
}